Create wrangler.toml
The core configuration file of Cloudflare Workers. This file tells Cloudflare's command-line tool wrangler how to build and deploy your project. It defines:

The name of the service and the main entry file.

The location of the static resource directory (i.e., the public directory).

The environment variables that need to be bound (such as ABOUT_INFO) and confidential information (such as GEMINI_API_KEY).