Update script.js
A more reliable proportion calculation method was adopted to correct this problem. It no longer relies on a fixed pixel size but calculates the index of the checkerboard grid based on the percentage of the click position on the canvas element, which can completely eliminate the influence brought by CSS scaling.
The relevant logic of loading the playback file and setting the interface (setupReplay) was reviewed and restructured to ensure that all HTML elements were ready before creating the button. Meanwhile, some log outputs have been added. If the file loading or parsing fails, more detailed error messages can be seen in the developer console of the browser (open by pressing F12).